Add separate modules for providing Retrofit instances
