+++ app/src/main/java/org/simple/clinic/di/NetworkModule.kt
@@ -5,7 +5,6 @@ import dagger.Module
 import dagger.Provides
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import org.simple.clinic.BuildConfig
 import org.simple.clinic.illustration.DayOfMonth
 import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.overdue.Appointment
@@ -24,14 +23,9 @@ import org.simple.clinic.util.moshi.LocalDateMoshiAdapter
 import org.simple.clinic.util.moshi.MoshiOptionalAdapterFactory
 import org.simple.clinic.util.moshi.URIMoshiAdapter
 import org.simple.clinic.util.moshi.UuidMoshiAdapter
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import retrofit2.converter.moshi.MoshiConverterFactory
-import retrofit2.converter.scalars.ScalarsConverterFactory
 import java.util.concurrent.TimeUnit
-import javax.inject.Named
 
-@Module(includes = [HttpInterceptorsModule::class])
+@Module(includes = [HttpInterceptorsModule::class, RetrofitModule::class])
 class NetworkModule {
 
   @Provides
@@ -80,38 +74,6 @@ class NetworkModule {
           // good number for syncing large batch sizes.
           readTimeout(configReader.long("networkmodule_read_timeout", default = 30L), TimeUnit.SECONDS)
         }
-        .build()
-  }
-
-  @Provides
-  @AppScope
-  fun retrofitBuilder(moshi: Moshi, okHttpClient: OkHttpClient): Retrofit.Builder {
-    return Retrofit.Builder()
-        .addConverterFactory(ScalarsConverterFactory.create())
-        .addConverterFactory(MoshiConverterFactory.create(moshi))
-        .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-        .client(okHttpClient)
-        .validateEagerly(true)
-  }
-
-  @Provides
-  @AppScope
-  @Named("for_country")
-  fun retrofit(commonRetrofitBuilder: Retrofit.Builder): Retrofit {
-    val baseUrl = BuildConfig.API_ENDPOINT
-    val currentApiVersion = "v3"
-
-    return commonRetrofitBuilder
-        .baseUrl("$baseUrl$currentApiVersion/")
-        .build()
-  }
-
-  @Provides
-  @AppScope
-  @Named("for_config")
-  fun configurationRetrofit(commonRetrofitBuilder: Retrofit.Builder): Retrofit {
-    return commonRetrofitBuilder
-        .baseUrl(BuildConfig.MANIFEST_ENDPOINT)
         .build()
   }
 }
